// src/app/core/models/media-type.enum.ts
/**
 * Перечисление для типов медиа-контента.
 * Используется для строгой типизации в коде.
 */
export enum MediaType {
  All = 'all',
  Movie = 'movie',
  Tv = 'tv',
}

/**
 * Перечисление для типов сортировки.
 */
export enum SortType {
  Newest = 'newest',
  TopRated = 'top_rated',
}